% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimalProg.R
\name{primal}
\alias{primal}
\title{(Generic) Primal Debiasing Program}
\usage{
primal(X, x, Psi, gamma = 0.1, quiet = TRUE)
}
\arguments{
\item{X}{Covariates.}

\item{x}{Querry point at which to evaluate the conditional quantile
function of Y given X = x.}

\item{Psi}{Estimate of the density matrix.}

\item{gamma}{Tuning parameter for primal debiasing program.}

\item{quiet}{If quiet = TRUE, then no messages are displayed.}
}
\value{
w    Solution to the primal debiasing program.
}
\description{
This function solves the primal debiasing program which yields MSE-optimal
weights to debias (any) Lasso pilot estimate.
}
